% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_db.R
\name{connect_to_db}
\alias{connect_to_db}
\title{Connect to a database}
\usage{
connect_to_db(url, url_env = "")
}
\arguments{
\item{url}{a specific URL}

\item{url_env}{URL to get from your .Renviron file}
}
\value{
connection
}
\description{
Given either a connection string (\code{url}) or the name of a URL to get from your .Renviron (\code{url_env}), return a connection to the database.
}
\seealso{
Other dbconnect: 
\code{\link{connect_to_soccer_db}()}
}
\concept{dbconnect}
