% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_all.R
\name{list_all_materialzied_views}
\alias{list_all_materialzied_views}
\title{List all materialized views in the database}
\usage{
list_all_materialzied_views(conn)
}
\arguments{
\item{conn}{a connection object}
}
\value{
tibble
}
\description{
Given a connection, return a tibble of all of the materialized views. (Canzhi's soccer db doesn't have any views, just materialized views.)
}
\seealso{
Other list_all: 
\code{\link{list_all_tables}()}
}
\concept{list_all}
