% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_all.R
\name{list_all_tables}
\alias{list_all_tables}
\title{List all tables in the database}
\usage{
list_all_tables(conn)
}
\arguments{
\item{conn}{a connection object}
}
\value{
tibble
}
\description{
Given a connection, return a tibble of all of the tables
}
\seealso{
Other list_all: 
\code{\link{list_all_materialzied_views}()}
}
\concept{list_all}
