% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_all.R
\name{pull_materialized_views}
\alias{pull_materialized_views}
\title{Get materialized view SQL}
\usage{
pull_materialized_views(name)
}
\arguments{
\item{views}{names of materialized views for which to pull sql}
}
\value{
character vector
}
\description{
Given a character vector of names (\code{views}), return the sql used to generate those materialized views
}
