% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_canzhi_db}
\alias{query_canzhi_db}
\title{Query a database}
\usage{
query_canzhi_db(conn, statement)
}
\arguments{
\item{conn}{active DB connection}

\item{statement}{sql}
}
\value{
tibble
}
\description{
Given a DB connection and an SQL statement, return a tibble with the result of the query.
}
