% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{connect_to_postgres_db}
\alias{connect_to_postgres_db}
\title{Connect to a Postgres Redshift database}
\usage{
connect_to_postgres_db(url)
}
\arguments{
\item{url}{a specific URL}
}
\value{
connection
}
\description{
Connect to a Postgres Redshift database
}
\seealso{
Other db: 
\code{\link{connect_to_understat_db}()}
}
\concept{db}
