% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_db.R
\name{connect_to_understat_db}
\alias{connect_to_understat_db}
\title{Connect to the soccer database}
\usage{
connect_to_understat_db()
}
\value{
connection
}
\description{
Connect to the soccer database
}
\seealso{
Other dbconnect: 
\code{\link{connect_to_postgres_db}()}
}
\concept{dbconnect}
