% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{connect_to_understat_db}
\alias{connect_to_understat_db}
\title{Connect to the understat database}
\usage{
connect_to_understat_db()
}
\value{
connection
}
\description{
Connect to the understat database
}
\seealso{
Other db: 
\code{\link{connect_to_postgres_db}()}
}
\concept{db}
